Moved NumpyArrayField to separate module
